handle cases when minifier is not available

+++ minify-concat.sh
@@ -9,5 +9,10 @@ do
   cat $file".css" >> master.css
 done < manifest
 
-minify master.css
-rm master.css
+if type minify > /dev/null 2>&1; then
+  minify master.css
+  rm master.css
+else
+  echo "\033[1;33m"Minify not found. Try 'npm install -g minifier'"\033[00m";
+  mv master.css master.min.css
+fi
